@@ -28,7 +28,7 @@
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 
         <!-- Java -->
-        <jdk.version>1.7</jdk.version>
+        <jdk.version>1.8</jdk.version>
 
         <!-- Maven Plugins -->
         <maven-compiler-plugin.version>3.1</maven-compiler-plugin.version>
